refactor(gmail): consolidate type imports for better clarity and organization
